refactoring. Виджеты вынесены в отдельные файлы
