+++ lib/ui/screen/visiting_screen.dart
@@ -1,10 +1,9 @@
 import 'dart:ui';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:places/domain/sight.dart';
 import 'package:places/mocks.dart';
-import 'package:places/resources/resources.dart';
-import 'package:places/ui/screen/sight_card.dart';
+import 'package:places/ui/widgets/visited_list_widget.dart';
+import 'package:places/ui/widgets/want_to_visit_list_widget.dart';
 import 'package:places/utils/app_color.dart';
 
 // Экран избранного с двумя таббарами. Реализовано через DefaultTabController
@@ -35,63 +34,6 @@ class _VisitingScreenState extends State<VisitingScreen> {
         ),
       ),
     );
-  }
-}
-
-//Виджет со списком посещенных мест
-class VisitedListWidget extends StatelessWidget {
-  final List<Sight> sightList;
-  const VisitedListWidget({Key? key, required this.sightList})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return sightList.isEmpty
-        ? PlaceholderVisitListWidget(
-            image: SvgPicture.asset(
-              Svgs.go,
-              width: 64,
-              height: 64,
-              color: AppColor.inactivGreyButtonColor,
-            ),
-            title: 'Пусто',
-            subtitle: 'Завершите маршрут,\nчтобы место попало сюда.',
-          )
-        : Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              children: [SightCard(sight: mocks[0])],
-            ),
-          );
-  }
-}
-
-// Виджет со списком мест которые хочу посетить
-class WantToVisitListWidget extends StatelessWidget {
-  final List<Sight> sightList;
-
-  const WantToVisitListWidget({Key? key, required this.sightList})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return sightList.isEmpty
-        ? PlaceholderVisitListWidget(
-            image: SvgPicture.asset(
-              Svgs.card,
-              width: 64,
-              height: 64,
-              color: AppColor.inactivGreyButtonColor,
-            ),
-            title: 'Пусто',
-            subtitle: 'Отмечайте понравившиеся\n места и они появиятся здесь.',
-          )
-        : Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              children: [SightCard(sight: mocks[0])],
-            ),
-          );
   }
 }
 
@@ -162,54 +104,5 @@ class AppBarFavorites extends StatelessWidget implements PreferredSizeWidget {
         ),
       ),
     );
-  }
-}
-
-// Верстка плейсхолдера для списков мест (если списки пусты)
-class PlaceholderVisitListWidget extends StatelessWidget {
-  final String title;
-  final String subtitle;
-  final Widget image;
-
-  const PlaceholderVisitListWidget({
-    Key? key,
-    required this.title,
-    required this.subtitle,
-    required this.image,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-        child: Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 53),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          image,
-          const SizedBox(height: 32),
-          Text(
-            title,
-            style: const TextStyle(
-              fontFamily: 'Roboto',
-              fontSize: 18,
-              fontWeight: FontWeight.w500,
-              color: AppColor.inactivGreyButtonColor,
-            ),
-          ),
-          const SizedBox(height: 8),
-          Text(
-            subtitle,
-            textAlign: TextAlign.center,
-            style: const TextStyle(
-              fontFamily: 'Roboto',
-              fontSize: 14,
-              fontWeight: FontWeight.w400,
-              color: AppColor.inactivGreyButtonColor,
-            ),
-          ),
-        ],
-      ),
-    ));
   }
 }
